fix(patterns): move the scoped-provenance search to state-inspector
The patterns package runs its plain Deno unit tests under a flat import
map, test-import-map.json, passed with --import-map. That map replaces
`commonfabric` with a small shim for every module in the graph, not just
for the pattern sources under test. It also carries the handful of bare
specifiers those sources need, and nothing else.

The agent sessions debug console added a provenance test that imports
`@commonfabric/state-inspector`. That drags the real runner, static and
js-compiler sources into the same graph. Those sources import specifiers
the flat map does not carry, such as `@commonfabric/ports`,
`@std/path/posix` and `turndown`, and they import `commonfabric` for the
real runtime API, which the shim does not define. Type checking reported
73 errors, so the patterns package failed in the workspace test job on
every push.

The failing case exercises no pattern code. It seeds a SQLite space in
which two identities write the same entity and both then overwrite their
first value, and checks that the scope overlay, the per-scope history
and the value reconstruction together recover the earlier snapshot. That
is a state-inspector test, and packages/state-inspector/test already
holds twenty tests that build a database the same way. Move it there as
scope-candidate-history.test.ts, alongside scopes.test.ts, which seeds
the same two tables.

The rest of the provenance tests check the shape of the command strings
the console prints for an operator to run. They exercise the pattern's
own module, need nothing but @std/assert, and stay where they are.

With the state-inspector import gone, the dependency entries the console
commit added to test-import-map.json have no remaining consumer, so this
restores that file to its previous contents.
