+++ packages/patterns/test-import-map.json
@@ -1,17 +1,11 @@
 {
   "imports": {
-    "@db/sqlite": "jsr:@db/sqlite@0.13.0",
-    "@noble/hashes": "npm:@noble/hashes@^2.2.0",
-    "@noble/hashes/": "npm:/@noble/hashes@^2.2.0/",
     "@std/assert": "jsr:@std/assert@^1",
     "@std/expect": "jsr:@std/expect@^1",
     "@std/jsonc": "jsr:@std/jsonc@^1",
     "@std/path": "jsr:@std/path@^1",
     "@std/testing/bdd": "jsr:@std/testing@^1/bdd",
     "@commonfabric/test-support/isolated-deno": "../test-support/src/isolated-deno.ts",
-    "commonfabric": "./tools/test-support/commonfabric.ts",
-    "hash-wasm": "npm:hash-wasm@^4.12.0",
-    "multiformats": "npm:multiformats@^14.0.4",
-    "multiformats/": "npm:/multiformats@^14.0.4/"
+    "commonfabric": "./tools/test-support/commonfabric.ts"
   }
 }
